finish 4.21

+++ controllers/blogs.js
@@ -49,8 +49,24 @@ blogsRouter.post('/api/blogs', async (request, response) => {
 })
 
 blogsRouter.delete('/api/blogs/:id', async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id)
-	response.status(204).end()
+	const blog = await Blog.findById(request.params.id)
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET) // token if have single line
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+	if (blog.user.toString() === user.id.toString()) {
+
+		await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).end()
+	}
+
+	else {
+		response.status(401).json({ error: 'this user cannot delete this blog!' })
+	}
+
+	
 })
 
 blogsRouter.put('/api/blogs/:id', async (request, response) => {
